feat: read the body once to make ABSOLUTELY sure the body doesn't exceed size limits
